backend/pyproject.toml
[tool.poetry]
name = "areg-packages"
version = "0.1.0"
description = ""
authors = ["Alg Reg"]

[tool.poetry.dependencies]
python = "^3.12"
SQLAlchemy = "^2.0.21"
python-dotenv = "^0.21.0"
alembic = "^1.8.1"
psycopg2-binary = "^2.9.5"
openpyxl = "^3.0.10"
python-slugify = "^7.0.0"
cryptography = "^41.0.7"
passlib = "^1.7.4"
wheel = "^0.38.1"
pytest = "^7.1.2"
coverage = "^6.0"
pyjwt = {version = "2.10.0", extras = ["crypto"]}
certifi = "^2023.5.7"
slowapi = "^0.1.8"
lxml = "^4.9.2"
pytest-mock = "^3.11.1"
python-docx = "^0.8.11"
flake8 = "^5.0.4"
black = {extras = ["jupyter"], version = "^22.10.0"}
pymupdf = "^1.23.2"
reportlab = "^4.0.4"
delayed_assert = "^0.3.6"
deepl = "^1.16.1"
pyotp = "^2.9.0"
fastapi-cache2 = "^0.2.1"
fastapi-utils = "^0.7.0"
beautifulsoup4 = "^4.12.3"
plotly = "^5.22.0"
kaleido = [
    # Newest version hangs on Windows, see https://github.com/plotly/Kaleido/issues/110
    {version = "0.1.0post1", markers = "platform_system == 'Windows'" },
    {version = "0.2.1", markers = "platform_system != 'Windows'"}
]
pandas = "^2.2.3"
pgvector = "^0.3.2"
snowballstemmer = "^2.2.0"
nltk = "^3.8.1"
transformers = "^4.43.4"
tensor = "^0.3.6"
torch = [
    {version = "2.6.0", markers = "platform_system == 'Windows'" },
    {url = "https://download.pytorch.org/whl/cpu/torch-2.6.0%2Bcpu-cp312-cp312-linux_x86_64.whl", markers = "platform_system != 'Windows'"}
]
numpy = "^1.25.2"
fastapi = "^0.115.6"
python-keycloak = "^5.1.1"
pydantic-settings = "^2.7.0"
thefuzz = {extras = ["speedup"], version = "^0.22.1"}

[tool.poetry.dev-dependencies]
ipykernel = "^6.16.2"

[tool.poetry.group.dev.dependencies]
black = {version = "^22.10.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
